Prodi

+++ app/Http/Controllers/MataKuliahController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\MataKuliah;
 use Illuminate\Http\Request;
 use App\ProgramStudi;
-use Illuminate\Support\Facades\Auth;
-
+Use App\User;
 class MataKuliahController extends Controller
 {
     /**
@@ -22,15 +22,18 @@ class MataKuliahController extends Controller
         ]);
     }
 
-    public function lihatMataKuliah()
+    public function lihatMataKuliah($program_studi_kode_prodi)
     {
-        $prodi=Auth::user()->program_studi_kode_prodi;
-        $kodeMataKuliah = MataKuliah::where('program_studi_kode_prodi', $prodi)->get();
-        $kodeMK = MataKuliah::where('program_studi_kode_prodi', $prodi)->value('id');
+        $user = Auth::user();
+        $programStudiKode = $user->program_studi_kode_prodi;
+        $kodeMataKuliah = MataKuliah::where('program_studi_kode_prodi', $program_studi_kode_prodi)->get(); 
+        $namaProgramStudi = ProgramStudi::where('kode_prodi', $programStudiKode)->value('nama'); 
+        $kodeMK = MataKuliah::where('program_studi_kode_prodi', $program_studi_kode_prodi)->value('id'); 
         return view('matkul.dashboard', [
             'kodeMataKuliahs' => $kodeMataKuliah,
-            'kodeProdi' => $prodi,
+            'kodeProdi' => $program_studi_kode_prodi,
             'kodeMK' => $kodeMK,
+            'namaProgramStudia' => $namaProgramStudi
         ]);
     }
 
@@ -40,7 +43,8 @@ class MataKuliahController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create($program_studi_kode_prodi)
-    {
+    {  
+        
         return view('matkul.create', [
             'kodeMataKuliaha' => $program_studi_kode_prodi,
         ]);
@@ -52,22 +56,9 @@ class MataKuliahController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request , $program_studi_kode_prodi)
+    public function store(Request $request , $program_studi_kode_prodi, $kodeMataKuliah)
     {
-        $validatedData = validator($request->all(),[
-            'kdMK' => 'required|string|max:100',
-            'txtName' => 'required|string|max:100',
-            'sks' => 'required|integer',
-            'semester' => 'required|integer'
-        ])->validate();
-        $mk = new MataKuliah();
-        $mk -> id = $validatedData['kdMK'];
-        $mk -> nama = $validatedData['txtName'];
-        $mk -> sks = $validatedData['sks'];
-        $mk -> semester = $validatedData['semester'];
-        $mk -> program_studi_kode_prodi = $program_studi_kode_prodi ;
-        $mk -> save();
-        return redirect(route('lihatMataKuliah',['kode'=>$program_studi_kode_prodi]));
+//
     }
 
     /**
@@ -87,9 +78,19 @@ class MataKuliahController extends Controller
      * @param  \App\MataKuliah  $mataKuliah
      * @return \Illuminate\Http\Response
      */
-    public function edit(MataKuliah $mataKuliah)
+    public function edit($program_studi_kode_prodi, $kodeMataKuliahi )
     {
-        //
+        $kodeMataKuliaha = MataKuliah::where('id', $kodeMataKuliahi)->where('program_studi_kode_prodi', $program_studi_kode_prodi)->first();   
+ 
+        if ($kodeMataKuliaha) {
+               return view('matkul.edit', [
+                'kodeMataKuliaha' => $kodeMataKuliaha,
+                'kodeMataKuliahs' => $program_studi_kode_prodi,
+                'idhaha'=>$kodeMataKuliahi,
+        ]);
+        }
+     
+        // dd($program_studi_kode_prodi, $kodeMataKuliahi,$kodeMataKuliaha );
     }
 
     /**
@@ -99,23 +100,41 @@ class MataKuliahController extends Controller
      * @param  \App\MataKuliah  $mataKuliah
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, MataKuliah $mataKuliah)
+    public function update(Request $request, $program_studi_kode_prodi, $kodeMataKuliahi, MataKuliah $mk)
     {
-        //
+        $validatedData = validator($request->all(), [
+            'kdMK' => 'required|string|max:100',
+            'txtName' => 'required|string|max:100',
+            'sks' => 'required|integer',
+            'semester' => 'required|integer'
+        ])->validate();
+    
+        $mk->where('id', $kodeMataKuliahi)->update([
+            'nama' => $validatedData['txtName'],
+            'sks' => $validatedData['sks'],
+            'semester' => $validatedData['semester'],
+            'program_studi_kode_prodi' => $program_studi_kode_prodi
+        ]);
+    
+        return redirect(route('lihatMataKuliah', ['kode' => $program_studi_kode_prodi]));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\MataKuliah  $id
+     * @param  \App\MataKuliah  $mataKuliah
      * @return \Illuminate\Http\Response
      */
-    public function destroy(MataKuliah $id): \Illuminate\Http\Response
+    public function destroy($program_studi_kode_prodi, $kodeMataKuliah)
     {
-        dd($id);
-//        $mataKuliah->id="IN455";
-//        dd($mataKuliah);
-        $mataKuliah->delete();
-        return redirect(route('lihatMataKuliah'));
+        // dd($program_studi_kode_prodi, $kodeMataKuliah);
+        $mataKuliah = MataKuliah::where('program_studi_kode_prodi', $program_studi_kode_prodi)
+            ->where('id', $kodeMataKuliah);
+    
+        if ($mataKuliah) {
+            $mataKuliah->delete();
+        }
+    
+        return redirect()->route('lihatMataKuliah', ['program_studi_kode_prodi' => $program_studi_kode_prodi]);
     }
 }
